requirements.txt
# ====================================================================================================
# Project Requirements — CustomPythonCoreFunctions v1.0
# ----------------------------------------------------------------------------------------------------
# Install all dependencies:
#     pip install -r requirements.txt
#
# Notes:
#   - tkinter is a built-in standard library and does NOT require installation.
#   - ttkbootstrap and tkcalendar are optional GUI enhancements.
#   - pdfminer.six is included because C00 imports extract_text directly.
#   - Only libraries used in C00 or G00a are listed here.
#   - No development-only tools included (e.g., pyinstaller).
# ====================================================================================================


# ----------------------------------------------------------------------------------------------------
# 1. CORE DATA & UTILITIES
# ----------------------------------------------------------------------------------------------------
pandas>=2.3
numpy>=1.26                  # Required because C00 imports numpy directly
tqdm>=4.66
requests>=2.32
pyyaml>=6.0.2


# ----------------------------------------------------------------------------------------------------
# 2. FILE & DOCUMENT HANDLING
# ----------------------------------------------------------------------------------------------------
pdfplumber>=0.11
PyPDF2>=3.0
pdfminer.six>=20231228
openpyxl>=3.1


# ----------------------------------------------------------------------------------------------------
# 3. DATABASE & CONNECTORS
# ----------------------------------------------------------------------------------------------------
snowflake-connector-python[secure-local-storage]>=4.0


# ----------------------------------------------------------------------------------------------------
# 4. WEB & AUTOMATION
# ----------------------------------------------------------------------------------------------------
selenium>=4.25
webdriver-manager>=4.0


# ----------------------------------------------------------------------------------------------------
# 5. GOOGLE API & AUTH
# ----------------------------------------------------------------------------------------------------
google-api-python-client>=2.100
google-auth>=2.40
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2


# ----------------------------------------------------------------------------------------------------
# 6. OPTIONAL GUI ENHANCEMENTS
# ----------------------------------------------------------------------------------------------------
ttkbootstrap>=1.10.1
tkcalendar>=1.6.1